transition_divisors implemented in AIR trait
